Modifications in preparation for harbour support
Rename files and references to them to have the "harbour-" prefix. Fixed .pro
file to some degree (redid indentation without tabs for consistency), with still
work to do to detect Sailfish in a way that Creator is aware (packagesExist is
not supported).

I still can't get a proper compilation and run on this crappy laptop, but this
should be good enough for now.